+++ packages/rum-core/src/domain/resource/resourceCollection.spec.ts
@@ -14,11 +14,11 @@ import { RumEventType } from '../../rawRumEvent.types'
 import type { RawRumEventCollectedData } from '../lifeCycle'
 import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
 import type { RequestCompleteEvent } from '../requestCollection'
-import { createTraceIdentifier } from '../tracing/tracer'
 import type { RumConfiguration } from '../configuration'
 import { validateAndBuildRumConfiguration } from '../configuration'
 import type { RumPerformanceEntry } from '../../browser/performanceObservable'
 import { RumPerformanceEntryType } from '../../browser/performanceObservable'
+import { createSpanIdentifier, createTraceIdentifier } from '../tracing/identifier'
 import { startResourceCollection } from './resourceCollection'
 
 const HANDLING_STACK_REGEX = /^Error: \n\s+at <anonymous> @/
@@ -65,6 +65,7 @@ describe('resourceCollection', () => {
       decodedBodySize: 51,
       transferSize: 63,
       renderBlockingStatus: 'blocking',
+      deliveryType: 'cache',
       responseStart: 250 as RelativeTime,
     })
     notifyPerformanceEntries([performanceEntry])
@@ -85,6 +86,7 @@ describe('resourceCollection', () => {
         first_byte: jasmine.any(Object),
         status_code: 200,
         protocol: 'HTTP/1.0',
+        delivery_type: 'cache',
         render_blocking_status: 'blocking',
       },
       type: RumEventType.RESOURCE,
@@ -122,6 +124,7 @@ describe('resourceCollection', () => {
         duration: (100 * 1e6) as ServerDuration,
         method: 'GET',
         status_code: 200,
+        delivery_type: undefined,
         protocol: undefined,
         type: ResourceType.XHR,
         url: 'https://resource.com/valid',
@@ -183,7 +186,7 @@ describe('resourceCollection', () => {
           createCompletedRequest({
             type: RequestType.XHR,
             traceId: createTraceIdentifier(),
-            spanId: createTraceIdentifier(),
+            spanId: createSpanIdentifier(),
             traceSampled: true,
           })
         )
@@ -236,6 +239,7 @@ describe('resourceCollection', () => {
         duration: (100 * 1e6) as ServerDuration,
         method: 'GET',
         status_code: 200,
+        delivery_type: undefined,
         protocol: undefined,
         type: ResourceType.FETCH,
         url: 'https://resource.com/valid',
@@ -312,7 +316,7 @@ describe('resourceCollection', () => {
         LifeCycleEventType.REQUEST_COMPLETED,
         createCompletedRequest({
           traceSampled: true,
-          spanId: createTraceIdentifier(),
+          spanId: createSpanIdentifier(),
           traceId: createTraceIdentifier(),
         })
       )
@@ -327,7 +331,7 @@ describe('resourceCollection', () => {
         LifeCycleEventType.REQUEST_COMPLETED,
         createCompletedRequest({
           traceSampled: false,
-          spanId: createTraceIdentifier(),
+          spanId: createSpanIdentifier(),
           traceId: createTraceIdentifier(),
         })
       )
@@ -348,7 +352,7 @@ describe('resourceCollection', () => {
         LifeCycleEventType.REQUEST_COMPLETED,
         createCompletedRequest({
           traceSampled: true,
-          spanId: createTraceIdentifier(),
+          spanId: createSpanIdentifier(),
           traceId: createTraceIdentifier(),
         })
       )
@@ -367,7 +371,7 @@ describe('resourceCollection', () => {
         LifeCycleEventType.REQUEST_COMPLETED,
         createCompletedRequest({
           traceSampled: true,
-          spanId: createTraceIdentifier(),
+          spanId: createSpanIdentifier(),
           traceId: createTraceIdentifier(),
         })
       )
@@ -387,7 +391,7 @@ describe('resourceCollection', () => {
         LifeCycleEventType.REQUEST_COMPLETED,
         createCompletedRequest({
           traceSampled: true,
-          spanId: createTraceIdentifier(),
+          spanId: createSpanIdentifier(),
           traceId: createTraceIdentifier(),
         })
       )